\name{ORDER2PARENT-package}
\alias{ORDER2PARENT-package}
\alias{ORDER2PARENT}
\docType{package}
\title{
Parent Distribution Estimation with Multiple Order Statistics
}
\description{
This package can estimate parent distribution nonparametrically when there are observations about multiple order statistics available. The estimator is based on B-spline.
}
\details{
\tabular{ll}{
Package: \tab ORDER2PARENT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
\code{\link{bgmm}}, \code{\link{blr}}, \code{\link{parentest}}
}
\author{
Cheng Chou

Maintainer: Cheng Chou <chengzho@usc.edu>
}
\keyword{ order statistic }
